+++ ixmp/core/timeseries.py
@@ -246,7 +246,28 @@ class TimeSeries:
 
     @property
     def url(self) -> str:
-        """URL fragment for the TimeSeries."""
+        """URL fragment for the TimeSeries.
+
+        This has the format ``{model name}/{scenario name}#{version}``, with the same
+        values passed when creating the TimeSeries instance.
+
+        Examples
+        --------
+        To form a complete URL (e.g. to use with :meth:`.from_url`), use a configured
+        :class:`Platform` name:
+
+        >>> platform_name = "my-ixmp-platform"
+        >>> mp = Platform(platform_name)
+        >>> ts = TimeSeries(mp, "foo", "bar", 34)
+        >>> ts.url
+        "foo/bar#34"
+        >>> f"ixmp://{platform_name}/{ts.url}"
+        "ixmp://platform_name/foo/bar#34"
+
+        .. note:: Use caution: because Platform configuration is system-specific, other
+           systems must have the same configuration for `platform_name` in order for
+           the URL to refer to the same TimeSeries/Scenario.
+        """
         return f"{self.model}/{self.scenario}#{self.version}"
 
     # Time series data